PBWDeformations

Documentation for PBWDeformations.

PBWDeformations.flatten

PBWDeformations.groupBy

PBWDeformations.flatten — Method

flatten(a::Vector{Vector{T}}) where T

Returns a vector of all elements of elements of a.

julia> flatten([[1],[],[2,3,4],[5],[]])
5-element Vector{Any}:
 1
 2
 3
 4
 5

PBWDeformations.groupBy — Method

groupBy(a::Vector{T}; eq=(==)) where T

Returns a vector containing the elements of a grouped into subvectors of consecutive equal elements.

julia> groupBy([1,1,2,2,2,2,3,1,4,4])
5-element Vector{Vector{Int64}}:
 [1, 1]
 [2, 2, 2, 2]
 [3]
 [1]
 [4, 4]

julia> groupBy([i for i in -5:5]; eq=((x, y) -> sign(x) == sign(y)))
3-element Vector{Vector{Int64}}:
 [-5, -4, -3, -2, -1]
 [0]
 [1, 2, 3, 4, 5]