Arc diagrams

Arc diagram induced bases

Warning

The basis ArcDiagDeformBasis can currently only be used for exterior powers of the standard module of Lie type $\mathfrak{so}_n$.

PBWDeformations.ArcDiagDeformBasis — Type

Concrete subtype of DeformBasis. Each element of the basis is induced by an arc diagram of a suitable size, which gets symmetrized and specialised to the given smash product. This process is due to Johannes Flake, Verity Mackscheidt (2022).

Reverse direction

Given a basis element of an above basis, there are functions to find one/all arc diagrams that induce it (up to a scalar). This is currently done by recomputing the basis and keeping track of the used arc diagrams.

PBWDeformations.corresponding_arc_diagram — Function

corresponding_arc_diagram(m::DeformationMap{C}, sp::SmashProductLie{C}, deg::Int) where {C <: RingElement}

Returns the arc diagram inducing the deformation map m of the smash product sp, where m is of the degree deg. If there is no such arc diagram, returns nothing.

corresponding_arc_diagram(m::DeformationMap{C}, sp::SmashProductLie{C}, degs::AbstractVector{Int}) where {C <: RingElement}

Returns the arc diagram inducing the deformation map m of the smash product sp, where m is of a degree in degs. If there is no such arc diagram, returns nothing.

PBWDeformations.corresponding_arc_diagrams — Function

corresponding_arc_diagrams(m::DeformationMap{C}, sp::SmashProductLie{C}, deg::Int) where {C <: RingElement}

Returns all arc diagrams of a suitable size inducing the deformation map m of the smash product sp, where m is of the degree deg. If there is no such arc diagram, returns an empty vector.

corresponding_arc_diagrams(m::DeformationMap{C}, sp::SmashProductLie{C}, degs::AbstractVector{Int}) where {C <: RingElement}

Returns all arc diagrams of a suitable size inducing the deformation map m of the smash product sp, where m is of a degree in degs. If there is no such arc diagram, returns an empty vector.