Smash products deformations

Constructors

PBWDeformations.DeformationMap — Type

DeformationMap{T} = MatElem{T} where {T <: SmashProductLieElem}

The type for deformation maps of a Lie algebra smash product. The entry kappa[i,j] should be the image of $v_i \wedge v_j$ under the deformation map, i.e. $κ(v_i,v_j)$. Deformation maps are always assumed to be quadratic and skew-symmetric.

PBWDeformations.deform — Function

deform(sp::SmashProductLie{C}, kappa::DeformationMap{elem_type(sp)}) where {C <: RingElem}

Constructs the deformation of the smash product sp by the deformation map kappa.

Returns a SmashProductLieDeform struct and a two-part basis.

PBWDeformations.symmetric_deformation — Function

symmetric_deformation(sp::SmashProductLie{C}) where {C <: RingElem}

Constructs the symmetric deformation of the smash product sp.

SmashProductLieDeform struct

PBWDeformations.SmashProductLieDeform — Type

The struct representing a deformation of a Lie algebra smash product. It consists of the underlying FreeAssociativeAlgebra with relations and some metadata. It gets created by calling deform.

Functions

The SmashProductLieDeform struct can be used as an argument for the following functions:

gen

gens

ngens

For gen, gens, and ngens, on can supply a symbol to choose the part of the smash product to use: :L for the Lie algebra, and :V for the module.