Smash products

Constructors

General / GAP case

PBWDeformations.smash_product_lie — Function

smash_product_lie(coeff_ring::Ring, symbL::Vector{Symbol}, symbV::Vector{Symbol}, struct_const_L, struct_const_V)

Constructs the smash product over the coefficient ring coeff_ring using the structure constants struct_const_L and struct_const_V, and using symbL and symbV as symbols for the respective generators of the lie algebra and the module.

smash_product_lie(coeff_ring::Ring, symbL::Vector{String}, symbV::Vector{String}, struct_const_L, struct_const_V)

The same as the other method with structure constants, but takes strings instead of symbols to name the generators.

smash_product_lie(coeff_ring::Ring, dynkin::Char, n::Int, lambda::Vector{Int})

Constructs the smash product of the abstract semisimple lie algebra given by dynkin and n and the highest weight module with weight lambda over the coefficient ring coeff_ring.

Example

julia> smash_product_lie(QQ, 'A', 1, [1])
(Lie Algebra Smash Product with basis x_1, x_2, x_3, v_1, v_2 over Rational Field, (QuadraticQuoAlgebraElem{fmpq}[x_1, x_2, x_3], QuadraticQuoAlgebraElem{fmpq}[v_1, v_2]))

SO case

For the orthogonal lie algebras so_n there is a different constructor, that results in the well-known basis of so_n given by x_i_j $= E_{i,j} - E_{j,i}$ for $1 \leq i < j \leq n$.

PBWDeformations.smash_product_lie_so — Function

smash_product_lie_so(coeff_ring :: Ring, n :: Int, lambda :: Vector{Int})

Constructs the smash product of the semisimple lie algebra so_n and the highest weight module with weight lambda over the coefficient ring coeff_ring. For n odd, lambda is a B-type highest weight; for n even, lambda is a D-type highest weight.

Example

julia> smash_product_lie_so(QQ, 5, [1,0])
(Lie Algebra Smash Product with basis x_1_2, x_1_3, x_1_4, ..., x_4_5, v_1, v_2, v_3, ..., v_5 over Rational Field, (QuadraticQuoAlgebraElem{fmpq}[x_1_2, x_1_3, x_1_4, x_1_5, x_2_3, x_2_4, x_2_5, x_3_4, x_3_5, x_4_5], QuadraticQuoAlgebraElem{fmpq}[v_1, v_2, v_3, v_4, v_5]))

Warning

Only fundamental non-spin representations are currently supported.

SmashProductLie struct

PBWDeformations.SmashProductLie — Type

The struct representing a lie algebra smash product. It consists of the underlying QuadraticQuoAlgebra and some metadata.

Functions

The SmashProductLie struct can be used as an argument for the following functions:

gens

ngens

change_base_ring