SmashProductPBWDeformLie functions

PBWDeformations.is_pbwdeform — Method

is_pbwdeform(d::SmashProductDeformLie{C}) where {C <: RingElement}

Check if d is a Poincare-Birkhoff-Witt deformation of a smash product. Uses Theorem 3.1 of Chelsea Walton, Sarah Witherspoon (2014).

PBWDeformations.pbwdeform_eqs — Method

pbwdeform_eqs(deform::SmashProductDeformLie{C}; disabled::Vector{Symbol}=Symbol[]) where {C <: RingElement}

Returns the equations for deform being a PBW deformation of a smash product as in Theorem 3.1 of Chelsea Walton, Sarah Witherspoon (2014). Subsets of the equations can be disabled by passing the corresponding symbols as keyword arguments, e.g. disabled = [:c, :d].

PBWDeformations.pbwdeforms_all — Method

pbwdeforms_all(sp::SmashProductLie{C}, degs::AbstractVector{Int}, DeformBasisType::Type{<:DeformBasis{C}}=DeformStdBasis{C}; special_return=Nothing) where {C <: RingElement}

Computes a basis of all Poincare-Birkhoff-Witt deformations of sp of degrees degs. DeformBasisType specifies the type of basis to use for the space of deformation maps. If special_return is SparseArrays.SparseMatrixCSC, the function returns intermediate results.

Uses pbwdeform_eqs and thus Theorem 3.1 of Chelsea Walton, Sarah Witherspoon (2014).

PBWDeformations.pbwdeforms_all — Method

pbwdeforms_all(sp::SmashProductLie{C}, deg::Int, DeformBasisType::Type{<:DeformBasis{C}}=DeformStdBasis{C}; special_return=Nothing) where {C <: RingElement}

The same as the other method, but only for a single degree deg.