Structure constants functions

To be able to use more sophisticated techniques, we need to have a description of Lie algebras in terms of matrices s.t. the standard representation operates by simple matrix-vector-multiplication. This is done by the functions in this section.

Since most of this package only considers structural constants of Lie algebras, modules etc. one needs to additionally save the knowledge about properties of the occurring objects and their bases, e.g. using SmashProductLieInfo for SmashProductLie.

PBWDeformations.liealgebra_so_basis — Method

liealgebra_so_basis(n::Int)

Returns a basis of the orthogonal Lie algebra of dimension n $\mathfrak{so}_n$.

It consists of the matrices $X_{ij} = E_{ij} - E_{ji}$ with $i < j$ in row-major order.

PBWDeformations.liealgebra_sp_basis — Method

liealgebra_sp_basis(n::Int)

Returns a basis of the symplectic Lie algebra $\mathfrak{sp}_{2n}$.

It consists of the following elements (where $i < j$):

$A_i = E_{i, i+n}$

$B_i = E_{i+n, i}$

$C_{ij} = E_{i, j} - E_{j+n, i+n}$

$D_{ij} = E_{j, i} - E_{i+n, j+n}$

$F_{ij} = E_{i, j+n} + E_{j, i+n}$

$G_{ij} = E_{i+n, j} + E_{j+n, i}$

$H_i = E_{i, i} - E_{i+n, i+n}$