Arc diagrams

Arc diagram induced bases

Warning

The basis ArcDiagDeformBasis can currently only be used for exterior and symmetric powers of the standard module of special orthogonal Lie algebras.

PBWDeformations.ArcDiagDeformBasis — Type

Concrete subtype of DeformBasis. Each element of the basis is induced by an arc diagram of a suitable size, which gets symmetrized and specialised to the given smash product. This process is due to [FM22].

Reverse direction

Given a basis element of an above basis, one can lookup all arc diagrams that induce it (up to a scalar).

PBWDeformations.lookup_params — Function

lookup_params(m::DeformationMap{T}, basis::ArcDiagBasedDeformBasis{ParamT, T}) where {T <: SmashProductLieElem}

Look up additional data that was used to generate the deformation map m in the basis basis. This can e.g. be an arc diagram or a pseudograph.