Smash products deformations

Constructors

PBWDeformations.DeformationMap — Type

DeformationMap{C} = Matrix{FreeAssAlgElem{C}} where {C <: RingElement}

The type for deformation maps of a Lie algebra smash product. The entry kappa[i,j] should be the image of $v_i \wedge v_j$ under the deformation map, i.e. $κ(v_i,v_j)$. Deformation maps are always assumed to be quadratic and skew-symmetric.

PBWDeformations.smash_product_deform_lie — Function

smash_product_deform_lie(sp::SmashProductLie{C}, kappa::DeformationMap{C}) where {C <: RingElement}

Constructs the deformation of the smash product sp by the deformation map kappa.

Returns a SmashProductDeformLie struct and a two-part basis.

PBWDeformations.smash_product_symmdeform_lie — Function

smash_product_symmdeform_lie(sp::SmashProductLie{C}) where {C <: RingElement}

Constructs the symmetric deformation of the smash product sp.

SmashProductDeformLie struct

PBWDeformations.SmashProductDeformLie — Type

The struct representing a deformation of a Lie algebra smash product. It consists of the underlying FreeAssAlgebra with relations and some metadata. It gets created by calling smash_product_deform_lie.

Functions

The SmashProductDeformLie struct can be used as an argument for the following functions:

gens

ngens

change_base_ring