Smash products

Constructors

General case

PBWDeformations.smash_product_lie — Function

smash_product_lie(coeff_ring::Ring, symbL::Vector{Symbol}, symbV::Vector{Symbol}, struct_const_L, struct_const_V)

Constructs the smash product over the coefficient ring coeff_ring using the structure constants struct_const_L and struct_const_V, and using symbL and symbV as symbols for the respective generators of the Lie algebra and the module.

Returns a SmashProductLie struct and a two-part basis.

smash_product_lie(coeff_ring::Ring, symbL::Vector{String}, symbV::Vector{String}, struct_const_L, struct_const_V)

The same as the other method with structure constants, but takes strings instead of symbols to name the generators.

Highest weight / GAP case

PBWDeformations.smash_product_lie_highest_weight — Function

smash_product_lie_highest_weight(coeff_ring::Ring, dynkin::Char, n::Int, lambda::Vector{Int})

Constructs the smash product of the abstract semisimple Lie algebra given by dynkin and n and the highest weight module with weight lambda over the coefficient ring coeff_ring.

Example

julia> smash_product_lie_highest_weight(QQ, 'A', 1, [1])
(Lie Algebra Smash Product with basis x_1, x_2, x_3, v_1, v_2 over Rational Field, (QuadraticQuoAlgebraElem{fmpq}[x_1, x_2, x_3], QuadraticQuoAlgebraElem{fmpq}[v_1, v_2]))

Constructive cases

$\mathfrak{so}_n$ – orthogonal Lie algebra

For the orthogonal Lie algebras $\mathfrak{so}_n$ there is a different constructor, that results in the well-known basis of $\mathfrak{so}_n$ given by x_i_j $= E_{i,j} - E_{j,i}$ for $1 \leq i < j \leq n$. See liealgebra_so_basis.

PBWDeformations.smash_product_lie_so_fundamental_module — Function

smash_product_lie_so_fundamental_module(coeff_ring::Ring, n::Int, e::Int)

Constructs the smash product of the Lie algebra $\mathfrak{so}_n$ and the e-th fundamental module over the coefficient ring coeff_ring.

PBWDeformations.smash_product_lie_so_symmpowers_standard_module — Function

smash_product_lie_so_symmpowers_fundamental_module(coeff_ring::Ring, n::Int, e::Int)

Constructs the smash product of the Lie algebra $\mathfrak{so}_n$ and the e-th symmetric power of the fundamental module over the coefficient ring coeff_ring.

PBWDeformations.smash_product_lie_so_extpowers_standard_module — Function

smash_product_lie_so_extpowers_standard_module(coeff_ring::Ring, n::Int, e::Int)

Constructs the smash product of the Lie algebra $\mathfrak{so}_n$ and the e-th exterior power of the fundamental module over the coefficient ring coeff_ring.

$\mathfrak{sp}_{2n}$ – symplectic Lie algebra

For more details about the basis used for $\mathfrak{sp}_{2n}$, refer to liealgebra_sp_basis.

PBWDeformations.smash_product_lie_sp_symmpowers_standard_module — Function

smash_product_lie_sp_symmpowers_standard_module(coeff_ring::Ring, n::Int, e::Int)

Constructs the smash product of the Lie algebra $\mathfrak{sp}_{2n}$ and the e-th symmetric power of the standard module over the coefficient ring coeff_ring.

PBWDeformations.smash_product_lie_sp_extpowers_standard_module — Function

smash_product_lie_sp_extpowers_standard_module(coeff_ring::Ring, n::Int, e::Int)

Constructs the smash product of the Lie algebra $\mathfrak{sp}_{2n}$ and the e-th exterior power of the fundamental module over the coefficient ring coeff_ring.

SmashProductLie struct

PBWDeformations.SmashProductLie — Type

The struct representing a Lie algebra smash product. It consists of the underlying QuadraticQuoAlgebra and some metadata. It gets created by calling smash_product_lie.

PBWDeformations.SmashProductLieInfo — Type

A struct containing additional information about a Lie algebra smash product. Every field can be set to nothing if it is unknown.

Name	Type	Description

dynkin	Char?	the family of the dynkin type of the Lie algebra

n	Int?	the n of the dynkin type of the Lie algebra

lambda	Vector{Int}?	highest weight vector of the module, only existing if the module is simple

constructive_basis	Bool	true if the used basis for the structure constants is known in terms of matrices

power_of_std_mod	Int?	if the module is a power of the standard module, positive = symmetric power, negative = exterior power

Functions

The SmashProductLie struct can be used as an argument for the following functions:

gens

ngens

change_base_ring