Smash products

Constructors

General case

PBWDeformations.smash_product — Function

smash_product(L::LieAlgebra{C}, V::LieAlgebraModule{C}) where {C <: RingElem}

Construct the smash product $TV \rtimes U(L)$.

smash_product(R::Ring, L::LieAlgebra{C}, V::LieAlgebraModule{C}) where {C <: RingElem}

Construct the smash product $TV \rtimes U(L)$ and extend the coefficients to R.

Highest weight / GAP case

SmashProductLie struct

PBWDeformations.SmashProductLie — Type

The struct representing a Lie algebra smash product. It consists of the underlying FreeAssAlgebra with relations and some metadata. It gets created by calling smash_product.

Functions

The SmashProductLie struct can be used as an argument for the following functions:

gen

gens

ngens

For gen, gens, and ngens, on can supply a symbol to choose the part of the smash product to use: :L for the Lie algebra, and :V for the module.