PBW Deformations of smash products

General deformation functions

PBWDeformations.is_pbwdeformation — Function

is_pbwdeformation(d::SmashProductLieDeform)

Check if d is a Poincare-Birkhoff-Witt deformation of a smash product. Uses Theorem 3.1 of [WW14].

PBWDeformations.pbwdeform_eqs — Function

pbwdeform_eqs(d::SmashProductLieDeform; disabled::Vector{Symbol}=Symbol[])

Returns the equations for d being a PBW deformation of a smash product as in Theorem 3.1 of [WW14]. Subsets of the equations can be disabled by passing the corresponding symbols as keyword arguments, e.g. disabled = [:c, :d].

All PBW deformations

PBWDeformations.all_pbwdeformations — Function

all_pbwdeformations(sp::SmashProductLie{C}, deform_basis::DeformBasis{C}; special_return=Nothing) where {C <: RingElem}

Computes a basis of all Poincare-Birkhoff-Witt deformations of sp. deform_basis specifies the basis to use for the space of deformation maps. If special_return is SMat, the function returns intermediate results.

Uses pbwdeform_eqs and thus Theorem 3.1 of [WW14].

all_pbwdeformations(sp::SmashProductLie{C}, degs::AbstractVector{Int}, DeformBasisType::Type{<:DeformBasis{C}}=StdDeformBasis{C}; special_return=Nothing) where {C <: RingElem}

Computes a basis of all Poincare-Birkhoff-Witt deformations of sp of degrees degs. DeformBasisType specifies the type of basis to use for the space of deformation maps. If special_return is SMat, the function returns intermediate results.

Uses pbwdeform_eqs and thus Theorem 3.1 of [WW14].

all_pbwdeformations(sp::SmashProductLie{C}, deg::Int, DeformBasisType::Type{<:DeformBasis{C}}=StdDeformBasis{C}; special_return=Nothing) where {C <: RingElem}

The same as the other method, but only for a single degree deg.

Bases of deformation map spaces

PBWDeformations.DeformBasis — Type

abstract type DeformBasis{C <: RingElem} end

A basis for a deformation map space of a Lie algebra smash product. The constructor of a subtype should accept a SmashProductLie and an AbstractVector{Int} of degrees. It is required that Base.length and Base.iterate are implemented for subtypes, where iterating yields objects of type DeformationMap{C}.

For a reference implementation, we refer to StdDeformBasis.

Standard basis

PBWDeformations.StdDeformBasis — Type

Concrete subtype of DeformBasis that implements the standard basis. Each element of the basis is a skew-symmetric matrix with 2 non-zero entries, where one entry is a pure tensor power of degree ∈ degs over the Lie algebra part of the smash product, and the other entry is its additive inverse.

Other bases

Please refer to Arc diagram induced bases and Pseudograph induced bases for more specialized bases.