Utility functions

This section contains some utility functions that are used in other parts of the package.

PBWDeformations.coefficient_vector — Method

coefficient_vector(M::MatElem{T}, basis::Vector{<:MatElem{T}}) where {T}

Returns the vector of coefficients of the matrix M in the basis basis. Requires that basis is linearly independent and that M lies in the span of basis.

Examples

julia> coefficient_vector(matrix(QQ, [1 2;3 4]), [matrix(QQ, [1 0;0 0]), matrix(QQ, [0 1;0 2]), matrix(QQ, [0 0;-1 0])])
[1   2   -3]

PBWDeformations.flatten — Method

flatten(a::Vector{Vector{T}}) where {T}

Returns a vector of all elements of elements of a.

Example

julia> flatten([[1],[],[2,3,4],[5],[]])
5-element Vector{Any}:
 1
 2
 3
 4
 5

PBWDeformations.groupBy — Method

groupBy(a::Vector{T}; eq=(==)) where {T}

Returns a vector containing the elements of a grouped into subvectors of consecutive equal elements.

Examples

julia> groupBy([1,1,2,2,2,2,3,1,4,4])
5-element Vector{Vector{Int64}}:
 [1, 1]
 [2, 2, 2, 2]
 [3]
 [1]
 [4, 4]

julia> groupBy([i for i in -5:5]; eq=((x, y) -> sign(x) == sign(y)))
3-element Vector{Vector{Int64}}:
 [-5, -4, -3, -2, -1]
 [0]
 [1, 2, 3, 4, 5]

PBWDeformations.is_valid_dynkin — Method

is_valid_dynkin(dynkin::Char, n::Int)

Returns true, if there given parameters uniquely define a dynkin diagram, i.e. are of one of the forms

$A_n$ for $n \geq 1$,

$B_n$ for $n \geq 2$,

$C_n$ for $n \geq 2$,

$D_n$ for $n \geq 4$,

$E_5$, $E_6$, $E_7$,

$F_4$,

$G_2$.

Examples

julia> is_valid_dynkin('A', 2)
true

julia> is_valid_dynkin('F', 4)
true

julia> is_valid_dynkin('D', 3)
false